[Gold IV] Title: 카드 정렬하기, Time: 320 ms, Memory: 25952 KB -BaekjoonHub

+++ 백준/Gold/1715. 카드 정렬하기/카드 정렬하기.java
@@ -7,16 +7,16 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         int n = Integer.parseInt(br.readLine());
 
-        PriorityQueue<Long> heap = new PriorityQueue<>();
+        PriorityQueue<Integer> heap = new PriorityQueue<>();
 
         for(int i = 0; i < n; i++) {
-            heap.offer(Long.parseLong(br.readLine()));
+            heap.offer(Integer.parseInt(br.readLine()));
         }
 
-        long result = 0;
+        int result = 0;
         while(heap.size() >= 2) {
-            long temp1 = heap.poll();
-            long temp2 = heap.poll();
+            int temp1 = heap.poll();
+            int temp2 = heap.poll();
 
             result += temp1 + temp2;
             heap.offer(temp1 + temp2);
